requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
Brotli==1.1.0
cachetools==6.2.1
celery==5.6.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
deep-translator==1.11.4
distro==1.9.0
Django==5.2.7
django-allauth==65.13.0
django-axes==8.0.0
django-cors-headers==4.9.0
django-csp==4.0
django-extensions==4.1
django-filter==25.2
django-formtools==2.5.1
django-htmx==1.26.0
django-otp==1.6.3
django-phonenumber-field==8.4.0
django-two-factor-auth==1.18.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
docx2txt==0.9
drf-spectacular==0.29.0
exceptiongroup==1.3.0
fonttools==4.60.1
freetype-py==2.5.1
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
inflection==0.5.1
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
lxml==6.0.2
numpy==2.2.6
openai==2.6.0
opencv-python==4.12.0.88
oscrypto==1.3.0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.11
psycopg-binary==3.2.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyparsing==3.2.5
pypdf==6.1.2
pypdfium2==5.2.0
pyphen==0.17.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rlPyCairo==0.4.0
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
pypdf==5.1.0
